+++ controller/AdminSAEController.class.php
@@ -69,66 +69,73 @@ class AdminSAEController extends Controller {
     }
 
     public function statistiquesAction(){
-        $id = $this->getRecuperateurGet()->getInt('id');
+        $authority_id = $this->getRecuperateurGet()->getInt('id');
+
+        $authoritySQL = $this->getObjectInstancier()->get(AuthoritySQL::class);
+        $authority_info = $authoritySQL->getInfo($authority_id);
+
+        $pastellPropertiesSQL = $this->getObjectInstancier()->get(PastellPropertiesSQL::class);
+        $pastellProperties = $pastellPropertiesSQL->getPastellProperties($authority_id);
 
 		$actesTransactionsSQL = $this->getObjectInstancier()->get(ActesTransactionsSQL::class);
-		$actesArchiveControler = $this->getObjectInstancier()->get(ActesArchiveControler::class);
 
 		$this->{'actes_nb_en_retard'} =
 			count($actesTransactionsSQL->getTransactionToArchive(
 				ActesPrepareSaeWorker::NB_DAYS_ARCHIVE_AFTER,
-				$id
+				$authority_id,
+				$pastellProperties->actes_send_auto
 			));
-		$this->{'actes_nb_en_attente_sae_4h'} =
+		$this->{'actes_nb_en_attente_transmission_sae'} =
 			$actesTransactionsSQL->getNbByStatusAndAuthority(
 				ActesStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE,
-				$id
+				$authority_id
 			);
 
-        $this->{'acte_nb_en_attente_auto'} =
-            $actesArchiveControler->getAllTransactionIdToSend();
-
-		$this->{'actes_nb_envoye_sae_4h'} =
+		$this->{'actes_nb_envoye_sae'} =
 			$actesTransactionsSQL->getNbByStatusAndAuthority(
 				ActesStatusSQL::STATUS_ENVOYE_AU_SAE,
-				$id
+				$authority_id
 			);
 		$this->{'actes_erreur_lors_de_larchivage'} =
 			$actesTransactionsSQL->getNbByStatusAndAuthority(
 				ActesStatusSQL::STATUS_ERREUR_LORS_DE_L_ARCHIVAGE,
-				$id
+				$authority_id
 			);
 		$this->{'actes_erreur_lors_de_lenvoi_sae'} =
 			$actesTransactionsSQL->getNbByStatusAndAuthority(
 				ActesStatusSQL::STATUS_ERREUR_LORS_DE_L_ENVOI_SAE,
-				$id
+				$authority_id
 			);
 
         $heliosTransactionsSQL = $this->getObjectInstancier()->get(HeliosTransactionsSQL::class);
 
-        $this->{'helios_nb_en_retard'} =
-            count($heliosTransactionsSQL->getTransactionToPrepareToSAE(
-                HeliosPrepareSaeWorker::NB_DAYS_ARCHIVE_AFTER,
-                $id
-            ));
-        $this->{'helios_nb_en_attente_sae_4h'} =
+		$this->{'helios_nb_en_retard'} =
+			count($heliosTransactionsSQL->getTransactionToPrepareToSAE(
+				HeliosPrepareSaeWorker::NB_DAYS_ARCHIVE_AFTER,
+				$authority_id,
+				$pastellProperties->helios_send_auto
+			));
+
+
+        $this->{'helios_nb_en_attente_transmission_sae'} =
             $heliosTransactionsSQL->getNbByStatusAndAuthority(
                 HeliosStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE,
-                $id
+                $authority_id
             );
-        $this->{'helios_nb_envoye_sae_4h'} =
+        $this->{'helios_nb_envoye_au_sae'} =
             $heliosTransactionsSQL->getNbByStatusAndAuthority(
 				HeliosStatusSQL::ENVOYER_AU_SAE,
-                $id
+                $authority_id
             );
         $this->{'helios_erreur_lors_de_lenvoi_sae'} =
             $heliosTransactionsSQL->getNbByStatusAndAuthority(
 				HeliosStatusSQL::STATUS_ERREUR_LORS_DE_L_ENVOI_SAE,
-                $id
+                $authority_id
             );
 
-        $this->{'authority_id'} = $id;
-        $this->title = "SAE - Statistiques sur une collectivité";
+        $this->{'authority_id'} = $authority_id;
+        $this->{'pastellProperties'} = $pastellProperties;
+        $this->title = "{$authority_info['name']} - Statistiques SAE";
     }
 
 }