+++ template/AdminSAEStatistiques.php
@@ -1,28 +1,33 @@
+<?php
+/** @var PastellProperties $pastellProperties */
 
-<h1>Statistiques envoi SAE</h1>
+?>
+<h1><?php hecho($title); ?></h1>
 <p id="back-transaction-btn">
     <a href="<?php echo WEBSITE_SSL ?>/admin/authorities/admin_authority_sae.php?id=<?php hecho($authority_id) ?>" class="btn btn-default">Retour configuration SAE</a>
 </p>
 
-<h2>Actes</h2>
+<h2>Actes <?php echo $pastellProperties->actes_send_auto?"(mode automatique)":""?></h2>
 
 <table class="data-table table table-striped ">
 
 
     <tr class="<?php echo $actes_nb_en_retard?"danger":"success" ?>">
-        <td>Actes en retard pour la transmission automatique</td>
+        <td>Actes à archiver (dans l'état acquittement reçu)</td>
         <td><span class="label label-<?php echo $actes_nb_en_retard?"danger":"success" ?>"><?php echo $actes_nb_en_retard ?></span></td>
         <td>
-            &nbsp;
+            <a href="/modules/actes/index.php?status=<?php echo ActesStatusSQL::STATUS_ACQUITTEMENT_RECU ?>&authority=<?php hecho($authority_id) ?>" class="icon">
+                Liste
+            </a>
         </td>
         <td>
             &nbsp;
         </td>
     </tr>
 
-    <tr class="<?php echo $actes_nb_en_attente_sae_4h?"danger":"success" ?>">
+    <tr class="<?php echo $actes_nb_en_attente_transmission_sae?"danger":"success" ?>">
         <td>Actes en attente de transmission au SAE</td>
-        <td><span class="label label-<?php echo $actes_nb_en_attente_sae_4h?"danger":"success" ?>"><?php echo $actes_nb_en_attente_sae_4h ?></span></td>
+        <td><span class="label label-<?php echo $actes_nb_en_attente_transmission_sae?"danger":"success" ?>"><?php echo $actes_nb_en_attente_transmission_sae ?></span></td>
         <td>
             <a href="/modules/actes/index.php?status=<?php echo ActesStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE ?>&authority=<?php hecho($authority_id) ?>" class="icon">
                 Liste
@@ -33,11 +38,9 @@
         </td>
     </tr>
 
-
-
-    <tr class="<?php echo $actes_nb_envoye_sae_4h?"danger":"success" ?>">
-        <td>Actes envoyé au SAE </td>
-        <td><span class="label label-<?php echo $actes_nb_envoye_sae_4h?"danger":"success" ?>"><?php echo $actes_nb_envoye_sae_4h ?></span></td>
+    <tr class="<?php echo $actes_nb_envoye_sae?"danger":"success" ?>">
+        <td>Actes en attente d'acceptation par le SAE</td>
+        <td><span class="label label-<?php echo $actes_nb_envoye_sae?"danger":"success" ?>"><?php echo $actes_nb_envoye_sae ?></span></td>
         <td>
             <a href="/modules/actes/index.php?status=<?php echo ActesStatusSQL::STATUS_ENVOYE_AU_SAE ?>&authority=<?php hecho($authority_id) ?>" class="icon">
                 Liste
@@ -57,7 +60,7 @@
             </a>
         </td>
         <td>
-            <form action="/modules/actes/actes_transac_change_status_sae_bulk.php" method="post" onsubmit="return confirm('Voulez-vous vraiment modifier ces transactions')">
+            <form action="/modules/actes/actes_transac_change_status_sae_bulk.php" method="post" onsubmit="return confirm('Voulez-vous vraiment modifier ces transactions ?')">
                 <input type="hidden" name="authority_id" value="<?php echo $authority_id ?>" />
                 <input type="hidden" name="status_id_from" value="<?php echo ActesStatusSQL::STATUS_ERREUR_LORS_DE_L_ENVOI_SAE ?>" />
                 <input type="hidden" name="status_id_to" value="<?php echo ActesStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE ?>" />
@@ -76,7 +79,7 @@
         </td>
         <td>
             <?php if($actes_erreur_lors_de_larchivage > 0) : ?>
-            <form action="/modules/actes/actes_transac_change_status_sae_bulk.php" method="post" onsubmit="return confirm('Voulez-vous vraiment modifier ces transactions')">
+            <form action="/modules/actes/actes_transac_change_status_sae_bulk.php" method="post" onsubmit="return confirm('Voulez-vous vraiment modifier ces transactions ?')">
                 <input type="hidden" name="authority_id" value="<?php echo $authority_id ?>" />
                 <input type="hidden" name="status_id_from" value="<?php echo ActesStatusSQL::STATUS_ERREUR_LORS_DE_L_ARCHIVAGE ?>" />
                 <input type="hidden" name="status_id_to" value="<?php echo ActesStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE ?>" />
@@ -90,36 +93,41 @@
 
 
 
-<h2>Helios</h2>
+<h2>Helios <?php echo $pastellProperties->helios_send_auto?"(mode automatique)":""?></h2>
 
 <table class="data-table table table-striped ">
 
 
     <tr class="<?php echo $helios_nb_en_retard?"danger":"success" ?>">
-        <td>Fichiers PES en retard pour la transmission automatique</td>
+        <td>Fichiers PES à archiver (dans l'état information disponible)</td>
         <td><span class="label label-<?php echo $helios_nb_en_retard?"danger":"success" ?>"><?php echo $helios_nb_en_retard ?></span></td>
         <td>
-            &nbsp;
+            <a href="/modules/helios/index.php?status=<?php echo HeliosStatusSQL::INFORMATION_DISPONIBLE ?>&authority=<?php hecho($authority_id) ?>" class="icon">
+                Liste
+            </a>
         </td>
+        <td>&nbsp;</td>
     </tr>
 
-    <tr class="<?php echo $helios_nb_en_attente_sae_4h?"danger":"success" ?>">
+    <tr class="<?php echo $helios_nb_en_attente_transmission_sae?"danger":"success" ?>">
         <td>Fichiers PES en attente de transmission au SAE</td>
-        <td><span class="label label-<?php echo $helios_nb_en_attente_sae_4h?"danger":"success" ?>"><?php echo $helios_nb_en_attente_sae_4h ?></span></td>
+        <td><span class="label label-<?php echo $helios_nb_en_attente_transmission_sae?"danger":"success" ?>"><?php echo $helios_nb_en_attente_transmission_sae ?></span></td>
         <td>
             <a href="/modules/helios/index.php?status=<?php echo HeliosStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE ?>&authority=<?php hecho($authority_id) ?>" class="icon">
                 Liste
             </a>
         </td>
+        <td>&nbsp;</td>
     </tr>
-    <tr class="<?php echo $helios_nb_envoye_sae_4h?"danger":"success" ?>">
-        <td>Fichiers PES envoyé au SAE </td>
-        <td><span class="label label-<?php echo $helios_nb_envoye_sae_4h?"danger":"success" ?>"><?php echo $helios_nb_envoye_sae_4h ?></span></td>
+    <tr class="<?php echo $helios_nb_envoye_au_sae?"danger":"success" ?>">
+        <td>Fichiers PES en attente d'acceptation par le SAE </td>
+        <td><span class="label label-<?php echo $helios_nb_envoye_au_sae?"danger":"success" ?>"><?php echo $helios_nb_envoye_au_sae ?></span></td>
         <td>
             <a href="/modules/helios/index.php?status=<?php echo HeliosStatusSQL::ENVOYER_AU_SAE ?>&authority=<?php hecho($authority_id) ?>" class="icon">
                 Liste
             </a>
         </td>
+        <td>&nbsp;</td>
     </tr>
 
     <tr class="<?php echo $helios_erreur_lors_de_lenvoi_sae?"danger":"success" ?>">
@@ -129,6 +137,14 @@
             <a href="/modules/helios/index.php?status=<?php echo HeliosStatusSQL::STATUS_ERREUR_LORS_DE_L_ENVOI_SAE ?>&authority=<?php hecho($authority_id) ?>" class="icon">
                 Liste
             </a>
+        </td>
+        <td>
+        <form action="/modules/helios/helios_transac_change_status_sae_bulk.php" method="post" onsubmit="return confirm('Voulez-vous vraiment modifier ces transactions ? ')">
+            <input type="hidden" name="authority_id" value="<?php echo $authority_id ?>" />
+            <input type="hidden" name="status_id_from" value="<?php echo HeliosStatusSQL::STATUS_ERREUR_LORS_DE_L_ENVOI_SAE ?>" />
+            <input type="hidden" name="status_id_to" value="<?php echo ActesStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE ?>" />
+            <input type="submit" class="btn btn-warning" value="Relancer" />
+        </form>
         </td>
     </tr>